Render at the output's framerate
Previously a moving pointer would schedule a frame when moving.
This means that if a pointer moves as fast as the output's frame
rate, frames will be sent at half of the output's frame rate.

Now, moving the pointer marks the output as dirty. When a frame
is rendered, a new frame is automatically scheduled. When the
compositor sends a frame event, we only render a frame if the
output is dirty.

Fixes #5