Fix off-by-one error in in_box function
Because the coordinate is 0 indexed, "greater than", not "greater or
equal than" has to be used here (imagine a 1x1 box, it only contains the
coordinates (0,0), not (1,1)).

This fixes output selection printing the wrong display name (as it uses
the upper left corner of the selection to figure out the display, it
falls into this edge case).