Fix incorrect anchor when output has negative position
When a selection is initiated from an output with a negative position,
anchor_x and anchor_y are appropriately set to negative values.

However, when subsequent pointer motion occurs, the removed code
was clamping the anchor values (which are later used as the top-left
corner of the selection) to be no less than zero.

Since outputs with negative positions will have negative anchor
values, let's remove this clamping.