Mark moved-away from outputs as dirty
Just like how pointer_handle_motion mark the old state as dirty,
pointer_handle_enter should also mark the old state as dirty.
This fixes #87.
Not sure whether leave or enter is preferred, but enter seems nicer
because that's closer to where the state is changed.

There is still a race condition left, although it's very unlikely to
happen:
1. handle_motion or handle_enter is called. The current selection is on output A.
2. The output A is marked dirty
3. render() is called async, and handles the redraw of A, using the old
   selection.
4. render() marks A a clean
5. move_seat is called from handle_motion or handle_enter. This updates
   the selection to point to output B and B is marked dirty
6. B will be redraw on further events, but A will forever remain in a