Print output local coordinates
The new format options '%X, %Y, %W, %H' will present the same property
as their lower case counterparts but are all with respect to the top
left corner of the selection. x and y coordinates are relative to the
top left corner of the output on which the selection starts.
If the selection spans over multiple outputs the width and hight will
be cropped to the intersection with the output.