Crosshairs (#95)
* Add support for crosshairs

* Refactor box functions into separate file

And have a separate header file too.

* Use cairo_translate to set cairo output coordinates to match global coordinates

This simplifies rendering, since we don't have to subtract the origin of
the output from the selection coordinates for each operation.

This also fixes the coordinates for the crosshairs drawn on the screen.
Previously, the calls to draw the crosshairs used a mix of both
coordinate systems, which would work for a single output, but not for
many multi-output setups.

* Use border color for crosshairs

The selection color defaults to being transparent, which is not very
visible if used over a light background, at least with the default
background color. I think using the border color is more useful.

* Add documentation for -x option

In man page and help output.

* Only mark outputs as dirty for crosshairs if necessary

---------

Co-authored-by: Jason Nader <jason.nader@protonmail.com>