format: Enforce format with clang-format
I tried to use config that minimized the changes, but there were some
existing inconsistencies in the format and some things I couldn't quite
get clang-format to due.

For example, I couldn't find a way to get clang-format to use more
indentation for continuation of parameter lists than for other types of
line continuation.

I also used ColumnLimit: 0, because anything else would add considerable
changes to either put things on the same line, split lines up, or both.

I figured if we want to set a limit, we could do that in a separate PR.

I'm also totally open to changing any of the configuration options.

If clang-format is too prescriptive, I could also just add something to
enforce that indentation is done with tabs instead of spaces.